+++ python/shipPatRec.py
@@ -7,7 +7,8 @@ import global_variables
 ReconstructibleMCTracks = []
 theTracks = []
 
-r_scale = global_variables.ShipGeo.strawtubes.InnerStrawDiameter / 1.975
+r_scale = 1.
+max_x = global_variables.ShipGeo.strawtubes.StrawLength # == 200 * u.cm
 
 def initialize(fgeo):
     pass
@@ -352,7 +353,7 @@ def fast_hough_pat_rec_stereo_views(SmearedHits_stereo, recognized_tracks_y, min
                 if ahit2['digiHit'] in used_hits:
                     continue
 
-                if abs(ahit1['zx_projection']) > 300 or abs(ahit2['zx_projection']) > 300:
+                if abs(ahit1['zx_projection']) > max_x or abs(ahit2['zx_projection']) > max_x:
                     continue
 
                 k_seed = 1. * (ahit2['zx_projection'] - ahit1['zx_projection']) / (ahit2['z'] - ahit1['z'])
@@ -369,7 +370,7 @@ def fast_hough_pat_rec_stereo_views(SmearedHits_stereo, recognized_tracks_y, min
                     if ahit3['digiHit'] in used_hits:
                         continue
 
-                    if abs(ahit3['zx_projection']) > 300:
+                    if abs(ahit3['zx_projection']) > max_x:
                         continue
 
                     layer3 = ahit3['detID'] // 10000
@@ -613,7 +614,7 @@ def artificial_retina_pat_rec_stereo_views(SmearedHits_stereo, recognized_tracks
         for ahit in SmearedHits_stereo:
             if ahit['digiHit'] in used_hits:
                 continue
-            if abs(ahit['zx_projection']) > 300:
+            if abs(ahit['zx_projection']) > max_x:
                 continue
             hits_z.append(ahit['z'])
             hits_x.append(ahit['zx_projection'])
@@ -636,7 +637,7 @@ def artificial_retina_pat_rec_stereo_views(SmearedHits_stereo, recognized_tracks
             if ahit3['digiHit'] in used_hits:
                 continue
 
-            if abs(ahit3['zx_projection']) > 300:
+            if abs(ahit3['zx_projection']) > max_x:
                 continue
 
             layer3 = ahit3['detID'] // 10000
@@ -911,7 +912,6 @@ def tracks_combination_using_extrapolation(recognized_tracks_12, recognized_trac
             atrack['hits_y34'] = recognized_tracks_34[i_34]['hits_y']
             atrack['hits_stereo34'] = recognized_tracks_34[i_34]['hits_stereo']
             recognized_tracks_combo.append(atrack)
-            recognized_tracks_combo.append(atrack)
 
     return recognized_tracks_combo
 
@@ -1021,7 +1021,7 @@ def pat_rec_stereo_views(SmearedHits_stereo, recognized_tracks_y, min_hits):
                 if ahit2['digiHit'] in used_hits:
                     continue
 
-                if abs(ahit1['zx_projection']) > 300 or abs(ahit2['zx_projection']) > 300:
+                if abs(ahit1['zx_projection']) > max_x or abs(ahit2['zx_projection']) > max_x:
                     continue
 
                 k_seed = 1. * (ahit2['zx_projection'] - ahit1['zx_projection']) / (ahit2['z'] - ahit1['z'])
@@ -1038,7 +1038,7 @@ def pat_rec_stereo_views(SmearedHits_stereo, recognized_tracks_y, min_hits):
                     if ahit3['digiHit'] in used_hits:
                         continue
 
-                    if abs(ahit3['zx_projection']) > 300:
+                    if abs(ahit3['zx_projection']) > max_x:
                         continue
 
                     layer3 = ahit3['detID'] // 10000